working on changes

+++ src/main/java/com/example/safaricomHackathon/repository/MovieRepository.java
@@ -3,5 +3,5 @@ package com.example.safaricomHackathon.repository;
 import com.example.safaricomHackathon.model.MovieModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MovieRepository extends CrudRepository<MovieModel, Long> {
+public interface MovieRepository extends CrudRepository<MovieModel, Integer> {
 }
